Split up counterexample implementation
